telegram_bot/requirements.txt
python-telegram-bot==21.0.1
openai==1.12.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
